requirements.txt
# Web 框架
Flask==3.1.2
flask-cors==6.0.2
flask-restx==1.3.2
flask-swagger-ui==5.21.0

# 数据库
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.45

# 加密与认证
cryptography==46.0.3
gmssl==3.2.2
Flask-JWT-Extended==4.7.1
PyJWT==2.10.1

# 工具
pytz==2024.1
